Extract Windows path syntax partial
